app/api/requirements.txt
Flask
psycopg2-binary
flask-cors
xrpl-py
requests
flask-sock
